#1 [setting] : move extension functions
