+++ app/src/main/java/com/example/market_kurly/core/util/context/ContextExt.kt
@@ -1,4 +1,4 @@
-package com.example.market_kurly.core.util
+package com.example.market_kurly.core.util.context
 
 import android.content.Context
 import android.widget.Toast
